backend/diarizer-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
torch==2.2.2
torchaudio==2.2.2
huggingface_hub==0.23.5
pyannote.audio==3.3.1
numpy>=1.26,<2
matplotlib>=3.8